DEV: Address review feedback
