--- assets/javascripts/discourse/templates/custom-topic-lists/header.hbs
@@ -1,1 +1,0 @@
-<h1>{{@model.title}}</h1>
